refactor: renamed benchmar dirs
