+++ src/rai_bench/rai_bench/tool_calling_agent/tasks/manipulation.py
